+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * `ws create` now supports `-i/-x` selectors to selectively update kapps
 * Defaults can now be set at the stack level (e.g. to pin helm/kubectl binaries per stack)
 * Sugarkube will now try to set up SSH port forwarding several times before failing
+* Breaking change: Replaced the 'skip' pre/post kapp action with 'none'. Conditions should now be used to control whether to install/delete kapps. The 'none' action can be used to prevent actions from running when overriding settings for kapps.
 
 ## 0.9.0
 * Rename the 'cache' subcommand to 'workspace' because it wasn't clear that users should actually work inside cache directories (i.e. they're not temporary caches).